% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_generics.R, R/SaemixData.R, R/SaemixRes.R,
%   R/SaemixObject.R, R/SaemixRes_c.R, R/SaemixRes_d.R
\docType{methods}
\name{initialize-methods}
\alias{initialize-methods}
\alias{initialize,SaemixData-method}
\alias{initialize,SaemixRepData-method}
\alias{initialize,SaemixSimData-method}
\alias{initialize,SaemixRes-method}
\alias{initialize,SaemixObject-method}
\alias{initialize,SaemixRes_c-method}
\alias{initialize,SaemixRes_d-method}
\title{Methods for Function initialize}
\usage{
\S4method{initialize}{SaemixData}(.Object, name.data, header, sep, na,
  name.group, name.predictors, name.response, name.covariates, name.X, units,
  name.mdv, name.cens, name.occ, name.ytype, verbose)

\S4method{initialize}{SaemixRepData}(.Object, data = NULL, nb.chains = 1)

\S4method{initialize}{SaemixSimData}(.Object, data = NULL, datasim = NULL)

\S4method{initialize}{SaemixRes}(.Object, modeltype, name.fixed, name.random,
  name.res, fixed.effects, fixed.psi, betaC, betas, omega, respar,
  cond.mean.phi, cond.var.phi, mean.phi, phi, phi.samp, parpop, allpar, MCOV)

\S4method{initialize}{SaemixObject}(.Object, data, model, options = list())

\S4method{initialize}{SaemixRes_c}(.Object, name.fixed, name.random, name.res,
  fixed.effects, fixed.psi, betaC, betas, omega, respar, cond.mean.phi,
  cond.var.phi, mean.phi, phi, phi.samp, parpop, allpar, MCOV)

\S4method{initialize}{SaemixRes_d}(.Object, name.fixed, name.random, name.res,
  fixed.effects, fixed.psi, betaC, betas, omega, cond.mean.phi, cond.var.phi,
  mean.phi, phi, phi.samp, parpop, allpar, MCOV)
}
\description{
Constructor functions for Classes in the saemix package
}
\section{Methods}{
 \describe{

\item{list("signature(.Object = \"SaemixData\")")}{ create a SaemixData
object. Please use the \code{\link{saemixData}} function.}

\item{list("signature(.Object = \"SaemixModel\")")}{ create a SaemixModel
object Please use the \code{\link{saemixModel}} function.}

\item{list("signature(.Object = \"SaemixObject\")")}{ create a SaemixObject
object. This object is obtained after a successful call to
\code{\link{saemix}}}

\item{list("signature(.Object = \"SaemixRepData\")")}{ create a
SaemixRepData object}

\item{list("signature(.Object = \"SaemixRes\")")}{ create a SaemixRes
object}

\item{list("signature(.Object = \"SaemixSimData\")")}{ create a
SaemixSimData object} }
}

\keyword{methods}
