% File src/library/parallel/man/parallel-package.Rd
% Part of the R package, http://www.R-project.org
% Copyright 2011 R Core Development Team
% Distributed under GPL 2 or later

\name{parallel-package}
\alias{parallel-package}
\alias{parallel}
\docType{package}
\title{
  Support for Parallel Computation.
}
\description{
  Support for parallel computation, including random-number generation.
}
\details{
  This package is under construction: it is intended to be released with
  \R 2.14.0.

  There is support for multiple RNG streams with the
  \samp{"L'Ecuyer-CMRG"} \link{RNG}: see \code{\link{nextRNGStream}}.
  
  For a complete list of functions, use \code{library(help="parallel")}.
}
\author{
  Maintainer: R Core Team \email{R-core@r-project.org}
}
\keyword{ package }
