% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaemixRes.R, R/SaemixRes_c.R, R/SaemixRes_d.R
\docType{class}
\name{SaemixRes-class}
\alias{SaemixRes-class}
\alias{SaemixRes}
\alias{SaemixRes-class}
\alias{SaemixRes}
\alias{SaemixRes-class}
\alias{SaemixRes}
\title{Class "SaemixRes"}
\description{
An object of the SaemixRes class, representing the results of a fit through the SAEM algorithm.

An object of the SaemixRes class, representing the results of a fit through the SAEM algorithm.

An object of the SaemixRes class, representing the results of a fit through the SAEM algorithm.
}
\section{Objects from the Class}{
 
An object of the SaemixRes class is returned as part of a SaemixObject object after a call to \code{\link{saemix}} and contain the following slots:
  \describe{
    \item{\code{name.fixed}:}{a vector containing the names of the fixed parameters in the model}
    }

 
An object of the SaemixRes class is returned as part of a SaemixObject object after a call to \code{\link{saemix}} and contain the following slots:
  \describe{
    \item{\code{name.fixed}:}{a vector containing the names of the fixed parameters in the model}
    }

 
An object of the SaemixRes class is returned as part of a SaemixObject object after a call to \code{\link{saemix}} and contain the following slots:
  \describe{
    \item{\code{name.fixed}:}{a vector containing the names of the fixed parameters in the model}
    }
}

\section{Methods}{

  \describe{
    \item{[<-}{\code{signature(x = "SaemixRes")}: replace elements of object}
    \item{[}{\code{signature(x = "SaemixRes")}: access elements of object}
    \item{initialize}{\code{signature(.Object = "SaemixRes")}: internal function to initialise object, not to be used}
    \item{print}{\code{signature(x = "SaemixRes")}: prints details about the object (more extensive than show)}
    \item{read}{\code{signature(object = "SaemixRes")}: internal function, not to be used }
    \item{showall}{\code{signature(object = "SaemixRes")}: shows all the elements in the object}
    \item{show}{\code{signature(object = "SaemixRes")}: prints details about the object}
    \item{summary}{\code{signature(object = "SaemixRes")}: summary of the results. Returns a list with a number of elements extracted from the results ().}
  }


  \describe{
    \item{[<-}{\code{signature(x = "SaemixRes_c")}: replace elements of object}
    \item{[}{\code{signature(x = "SaemixRes_c")}: access elements of object}
    \item{initialize}{\code{signature(.Object = "SaemixRes_c")}: internal function to initialise object, not to be used}
    \item{print}{\code{signature(x = "SaemixRes_c")}: prints details about the object (more extensive than show)}
    \item{read}{\code{signature(object = "SaemixRes_c")}: internal function, not to be used }
    \item{showall}{\code{signature(object = "SaemixRes_c")}: shows all the elements in the object}
    \item{show}{\code{signature(object = "SaemixRes_c")}: prints details about the object}
    \item{summary}{\code{signature(object = "SaemixRes_c")}: summary of the results. Returns a list with a number of elements extracted from the results ().}
	 }


  \describe{
    \item{[<-}{\code{signature(x = "SaemixRes_d")}: replace elements of object}
    \item{[}{\code{signature(x = "SaemixRes_d")}: access elements of object}
    \item{initialize}{\code{signature(.Object = "SaemixRes_d")}: internal function to initialise object, not to be used}
    \item{print}{\code{signature(x = "SaemixRes_d")}: prints details about the object (more extensive than show)}
    \item{read}{\code{signature(object = "SaemixRes_d")}: internal function, not to be used }
    \item{showall}{\code{signature(object = "SaemixRes_d")}: shows all the elements in the object}
    \item{show}{\code{signature(object = "SaemixRes_d")}: prints details about the object}
    \item{summary}{\code{signature(object = "SaemixRes_d")}: summary of the results. Returns a list with a number of elements extracted from the results ().}
	 }
}

\examples{
methods(class="SaemixRes")

showClass("SaemixRes")

methods(class="SaemixRes_c")

showClass("SaemixRes_c")

methods(class="SaemixRes_d")

showClass("SaemixRes_d")

}
\references{
Kuhn E, Lavielle M. Maximum likelihood estimation in nonlinear mixed effects models. Computational Statistics and Data Analysis 49, 4 (2005), 1020-1038.

Comets E, Lavenu A, Lavielle M. SAEMIX, an R version of the SAEM algorithm. 20th meeting of the Population Approach Group in Europe, Athens, Greece (2011), Abstr 2173.

Kuhn E, Lavielle M. Maximum likelihood estimation in nonlinear mixed effects models. Computational Statistics and Data Analysis 49, 4 (2005), 1020-1038.

Comets E, Lavenu A, Lavielle M. SAEMIX, an R version of the SAEM algorithm. 20th meeting of the Population Approach Group in Europe, Athens, Greece (2011), Abstr 2173.

Kuhn E, Lavielle M. Maximum likelihood estimation in nonlinear mixed effects models. Computational Statistics and Data Analysis 49, 4 (2005), 1020-1038.

Comets E, Lavenu A, Lavielle M. SAEMIX, an R version of the SAEM algorithm. 20th meeting of the Population Approach Group in Europe, Athens, Greece (2011), Abstr 2173.
}
\seealso{
\code{\link{saemixData}} \code{\link{SaemixModel}} \code{\link{saemixControl}} \code{\link{saemix}}

\code{\link{saemixData}} \code{\link{SaemixModel}} \code{\link{saemixControl}} \code{\link{saemix}}

\code{\link{saemixData}} \code{\link{SaemixModel}} \code{\link{saemixControl}} \code{\link{saemix}}
}
\author{
Emmanuelle Comets \email{emmanuelle.comets@inserm.fr}

Audrey Lavenu

Marc Lavielle.

Emmanuelle Comets \email{emmanuelle.comets@inserm.fr}

Audrey Lavenu

Marc Lavielle.

Emmanuelle Comets \email{emmanuelle.comets@inserm.fr}

Audrey Lavenu

Marc Lavielle.
}
\keyword{classes}
